server/requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
python-dotenv==1.0.0
pymupdf
psycopg2==2.9.6
langchain-openai==0.0.3
langchain==0.0.325
python-multipart
gunicorn
websockets